webapp/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-jose[cryptography]
google-auth
google-auth-oauthlib
httpx
